2requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
PyJWT
django-cors-headers